// src/photos/dto/create-session.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsUUID } from 'class-validator';

export class CreateSessionDto {
  @ApiProperty({
    description: 'UUID v4 generated by the photobooth',
    example: 'b5f4a6e2-3c1d-4f8a-9b1e-1a2b3c4d5e6f',
  })
  @IsUUID('4')
  sessionToken!: string;

  @ApiProperty({
    description: 'Event token UUID',
    example: '1487f80e-78db-426b-81de-ef56b1e62af8',
  })
  @IsUUID('4')
  eventToken!: string;
}